add Navigation Page model
