+++ runner_tests.sh
@@ -13,9 +13,11 @@ echo $LOGIN $PASSWORD $NICKNAME
 # pytest hw/code/test_film_unauth.py
 # pytest hw/code/test_film_auth.py
 
-# pytest hw/code/test_navigation.py
+pytest hw/code/test_navigation_auth.py
+# pytest hw/code/test_navigation_unauth.py
+
 # pytest hw/code/test_mainpage.py
 # pytest hw/code/test_premiere.py
 
-pytest hw/code/test_login.py
+# pytest hw/code/test_login.py
 # pytest hw/code/test_signup.py
